Refactor things a bit, NFC: typeCheckExpression takes a convertType and a contextualType, but the only client of contextualType is pattern binding initialization that wants to provide a hint to the initializer, but the type may not be fully formed (e.g. just Array instead of Array<Int>) so it can't provide a full convertType to that type.
After trying various ways of eliminating the concept and failing, repaint it
instead:  instead of typeCheckExpression taking *both* a convertType and a
contextualType (even though they are mutually exclusive), have it only take
a single convertType and add a bit to TypeCheckExprOptions to indicate whether
it is a hint or a strong conversion.  This simplifies all the callers of
typeCheckExpression and makes further evolution simpler.



Swift SVN r30658