[ls-opts] Make sure to invalidate StoreMap stores when we fail to eliminate a dead store.
Previously we were just eliminating stores from the StoreList, not the storemap.
This fixes it so that we properly elimiante them.

This bug has been in the compiler since the storemap code was added a long time
ago. The reason it was found now is that I tightened up the verification in load
store opts recently.

rdar://21952712

Swift SVN r30785