Add BuiltinInst::{getIntrinsicID(),getBuiltinKind()}.
Often times one will want to process a BuiltinInst just based off of the
intrinsic or builtin kind. Generally one only wants to process one or
the other and has to introduce variables for the IntrinsicInfo or
BuiltinInfo to a wider scope than one wishes to. These builtins simplify
the processing of such builtins by allowing one to create an if
statement scope around the variable.

if (auto Kind = BI->getBuiltinKind()) {
  ...
}

if (auto ID = BI->getIntrinsicID()) {
  ...
}

Swift SVN r30835