[sil-combine] Be more careful when removing convert_function instructions.
If the operand of a convert_function is not a simple function_ref, we need to replace any ref-counting instructions on convert_function by equivalent ref_counting instructions on its argument. Doing so preserves the life-time of the convert_function argument, which is especially important if it is a partial_apply.

Fixes rdar://22102172

Swift SVN r30944