Reimplement materializeForSet emission in SILGen instead of the type-checker.  The strategy for now is to just use this for protocol witness thunk emission, where it is required when generating a materializeForSet for storage that is either implemented in a protocol extension or requires reabstraction to the requirement's pattern.
Eventually, this should be generalized to the point that
we can use it for all materializeForSet emission, which
I don't think will take much.  However, that's not really
the sort of instability we want to embrace for the current
release.

WIP towards rdar://21836671; currently disabled, so NFC.

Swift SVN r31072