Remove the stand-alone devirtualizer pass.
It is no longer necessary since we devirtualize in both the mandatory
inliner and performance inliner.

I opened rdar://problem/22201243 so that we can investigate folding some
functionality into the inliner to avoid having to run the series of
passes required in devirt_try_apply.sil to fully devirtualize one of the
test cases.

Swift SVN r31095