More changes for specializing partial applies during inlining.
It turns out r31088 didn't have all the required changes to match what
happens in the stand-alone specialization pass (which will be removed in
a forthcoming change). We also need to collect all the new partial
applies in the specializations we create, since we may also specialize
the functions they partially apply.

Swift SVN r31111