When CSApply applies as Builtin type to a literal, use shallowClone to allocate a new node, abandoning the old node.  This cuts of places where we'd end up retypechecking the original expr and getting things like Builtin.Int2048 into the diagnostic.
This is NFC, because there is some crufty stuff avoiding this sitation in CSDiags that
needs to be removed for other reasons, this simply unblocks that.

I have a more aggressive approach in mind to actually fix how we model literals, but 
that isn't on my short-term todo list.



Swift SVN r31270