[sil-prespecialization] Whitelist the  Swift._allocateUninitializedArray
When pre-specialization was introduced I forgot to whitelist the _allocateUninitializedArray.

As a result, the creation of an array with elements of builtin types was using a generic code. Moreover, due to use of generics, a lot of getGenericMetadata runtime calls were performed, resulting in increased memory usage.

With this change, the memory usage of a "hello, world" Swift-app is down to 5.7KB.

Swift SVN r31322