Remove the stand-alone generic specialization pass.
We do generic specialization in the inliner now, so we no longer need
the separate pass.

The order the specializations are inserted into the module differs as a
result of this, which means that one unfortunately huge test needed
substantial changes.

At some point this test needs to be converted to SIL and split up into
smaller chunks.

Swift SVN r31323