[cmake] Add an option SWIFT_EXPERIMENTAL_EXTRA_NEGATIVE_REGEXP_FLAGS
This is a negative version of the flag
SWIFT_EXPERIMENTAL_EXTRA_REGEXP_FLAGS. The reason why an additional flag
is needed is that cmake does not support negative regexes in the regex
string itself. Instead you need to do:

if (NOT "..." MATCHES "...")
  ...
endif()

This suggests that we either needed to introduce more complexity into
SWIFT_EXPERIMENTAL_EXTRA_REGEXP_FLAGS or just introduce another flag. I
decided to go with the later.

Swift SVN r31417