Add Analysis.def to allow for metaprogramming with Analysis names.
This allows one by editing one place to create an Analysis kind and ensure that
your analysis is properly registered in all of the relevant places for use with
swift/sil-opt.

If we add tools like a sil bugpoint, this will make it easy to add analysis
without having to know about all of these locations.

It also standardizes the create***Analysis API to take a SILModule and a
SILPassManager. All passes take in a SILModule and some take in a
SILPassManager. By standardizing the API and in the cases where SILPassManager
is not used, just making it a dead argument enables metaprogramming.

Swift SVN r31420