stdlib: Cache the unicode collation values for the ASCII range and use this cache instead.
Create two tables. One that caches the ASCII ordering as returned by the root
collator. This table is static and we can use this for comparison and ordering.
The second table is cached the first time it is queried. This table caches the
actual collation values. We use this table for hashing. Collation values might
change between versions of the ICU dylib and so we can't store this in a static
table.

Use the second table to create a unicode hash for ascii strings.

The first table will be used in a follow-up commit.

Swift SVN r31472