Fix a test case that relied on a specific behavior of the optimizer.
The test unimplemented_initializer.swift checks that we don't leak file names
through calls to runtime trap function. Before this commit the test checked for
a specific sil function mame with specific attributes. The test expected that
the optimizer would preserve the function declaration, but "Function Signature
Optimization" changed the function signature and turned the original function
into a thunk. I removed the checks for a specific function name and kept the
CHECK and CHECK-NOT lines that preserve the original intent of the test.

Swift SVN r31521