Remove the old fuzzer script.
We've used this utility to generate piles of code to stress test SSA
construction and the other optimziation passes we had at that time, and also for
measuring the compile time of the passes.  But this little hack is no longer
maintained and it does not even generate legal swift any more so it is time to
delete it.

Swift SVN r31548