Fix an overactive assertion: <rdar://problem/22470302> Crash with parenthesized call result
We type check expressions using a contextual purpose of CTP_CalleeResult
without a specific contextualType, because we install the contextual type
as a conversion constraint.  This formerly failed the assertion expecting
that you have to have a type if you have a purpose, because parenexprs 
propagated their contextual info down.

In addition to making the assertion in TypeCheckConstraints.cpp more
lenient, change visitCallExpr to just pass down the purpose directly
instead of installing it in its ExprTypeCheckListener.



Swift SVN r31575